Function cli::commands::get_commands

pub fn get_commands(
    include_dev: bool
) -> (Vec<Arc<dyn Command>>, HashMap<&'static str, Arc<dyn Command>>)

Returns all the commands available, as well as the reverse index from the aliases to the commands.