cli

Enum CommandOptions

pub(crate) enum CommandOptions {

    Play {
        index: Option<usize>,
    },
    Pause,
    Stop,
    Next,
    Prev,
    Seek {
        seconds: i32,
    },
    Move {
        index1: usize,
        index2: usize,
    },
    Remove {
        index: Option<usize>,
    },
    Clear,
    List,
    Add {
        filenames: Vec<String>,
    },
    Replace {
        filenames: Vec<String>,
    },
    Prop {
        properties: Vec<String>,
    },
    Interactive,

}

Variants

Play

Play media at the specified index in the playlist

Fields

index: Option<usize>

The index of the media to play

Pause

Pause the currently playing media

Stop

Stop the playback and quit MPV

Next

Skip to the next item in the playlist

Prev

Skip to the previous item in the playlist

Seek

Seek to a specific position in the currently playing media

Fields

seconds: i32

The number of seconds to seek to

Move

Move an item in the playlist from one index to another

Fields

index1: usize

The index of the item to move

index2: usize

The index to move the item to

Remove

Remove an item from the playlist

If invoked while playlist has no entries, or if the only entry is the active video, then this will exit MPV.

Fields

index: Option<usize>

The index of the item to remove (optional)

Clear

Clear the entire playlist

List

List all the items in the playlist

Add

Add files to the playlist

Needs at least one file to be passed.

Fields

filenames: Vec<String>

The filenames of the files to add

Replace

Replace the current playlist with new files

Fields

filenames: Vec<String>

The filenames of the files to replace the playlist with

Prop

Fetch properties of the current playback or playlist

Fields

properties: Vec<String>

The properties to fetch

Interactive

Enter interactive mode to send commands to MPV IPC