Function cli::exec_path

pub fn exec_path() -> Path

The file path of the executed program.

Typically the file path of the invoked executable. If the invocation target is a symlink, then it will be resolved to the executable it links to.