cli

Module meta_snapshot

Structs

MetaConnection 🔒

MetaRestoreCommand

Restore metadata snapshot tool. This tool is used to restore metadata snapshot from etcd, pg or mysql. It will restore the metadata snapshot from local file or s3 bucket.

MetaRestoreTool

MetaSnapshotCommand

Export metadata snapshot tool. This tool is used to export metadata snapshot from etcd, pg or mysql. It will dump the metadata snapshot to local file or s3 bucket. The snapshot file will be in binary format.

MetaSnapshotTool

S3Config 🔒

Functions

create_local_file_object_store 🔒