Struct cli::Style

pub struct Style {
    // some fields omitted
}

Styles a text with ANSI styles. This styling is self contained which means that at the end of the string a reset code is issued. Examples:

use cli::{Style, Red};

let mut text = Style::new(String::from_str("hello"));
text.fg(Red);
println!("{}", text);

Methods

impl Style

fn new(text: String) -> Style

Create one new styled string with ansi codes.

fn fg(&mut self, color: Color)

Sets the foreground color.

fn bg(&mut self, color: Color)

Sets the background color.

fn bold(&mut self, bold: bool)

Enable or disable bold mode.

fn dim(&mut self, dim: bool)

Enable or disable dim mode.

fn underline(&mut self, underline: bool)

Enable or disable underline.

fn blink(&mut self, blink: bool)

Enable or disable blinking.

fn reverse(&mut self, reverse: bool)

Enable or disable inverse rendering.