Module cli::sysexitsExperimental

A collection of predefined exit codes cribbed from sysexits.h.

From sysexits.h:

According to style(9), it is not a good practice to call exit(3) with arbitrary values to indicate a failure condition when ending a program. Instead, the pre-defined exit codes from sysexits should be used, so the caller of the process can get a rough estimation about the failure class without looking up the source code.

Intended for use with std::os::set_exit_status prior to process exit.

Example

ignore use std::os; fn main() { // ... os::set_exit_status(cli::sysexits::USAGE); return; }

Constants

CANT_CREAT
Can't create (user) output file

CONFIG
Configuration error

DATA_ERR
Data format error

IO_ERR
Input/output error

NO_HOST
Host name unknown

NO_INPUT
Cannot open input

NO_PERM
Permission denied

NO_USER
Addressee unknown

OK
Successful termination

OS_ERR
System error (e.g. can't fork)

OS_FILE
Critical OS file missing

PROTOCOL
Remote error in protocol

SOFTWARE_ERR
Internal software error

TEMP_FAIL
Temp failure; user is invited to retry

UNAVAILABLE
Service unavailable

USAGE
Command line usage error