Struct cli::types::PlatformOverrideTask

pub struct PlatformOverrideTask {
    pub clear: Option<bool>,
    pub disabled: Option<bool>,
    pub private: Option<bool>,
    pub deprecated: Option<DeprecationInfo>,
    pub extend: Option<String>,
    pub watch: Option<TaskWatchOptions>,
    pub condition: Option<TaskCondition>,
    pub condition_script: Option<Vec<String>>,
    pub ignore_errors: Option<bool>,
    pub force: Option<bool>,
    pub env_files: Option<Vec<EnvFile>>,
    pub env: Option<IndexMap<String, EnvValue>>,
    pub cwd: Option<String>,
    pub install_crate: Option<InstallCrate>,
    pub install_crate_args: Option<Vec<String>>,
    pub install_script: Option<Vec<String>>,
    pub command: Option<String>,
    pub args: Option<Vec<String>>,
    pub script: Option<ScriptValue>,
    pub script_runner: Option<String>,
    pub script_extension: Option<String>,
    pub run_task: Option<RunTaskInfo>,
    pub dependencies: Option<Vec<String>>,
    pub toolchain: Option<String>,
}

Holds a single task configuration for a specific platform as an override of another task

Fields
clear: Option<bool>

if true, it should ignore all data in base task

disabled: Option<bool>

if true, the command/script of this task will not be invoked, dependencies however will be

private: Option<bool>

if true, the task is hidden from the list of available tasks and also cannot be invoked directly from cli

deprecated: Option<DeprecationInfo>

if not false, this task is defined as deprecated

extend: Option<String>

Extend any task based on the defined name

watch: Option<TaskWatchOptions>

set to true to watch for file changes and invoke the task operation

condition: Option<TaskCondition>

if provided all condition values must be met in order for the task to be invoked (will not stop dependencies)

condition_script: Option<Vec<String>>

if script exit code is not 0, the command/script of this task will not be invoked, dependencies however will be

ignore_errors: Option<bool>

if true, any error while executing the task will be printed but will not break the build

force: Option<bool>

DEPRECATED, replaced with ignore_errors

env_files: Option<Vec<EnvFile>>

The env files to setup before running the task commands

env: Option<IndexMap<String, EnvValue>>

The env vars to setup before running the task commands

cwd: Option<String>

The working directory for the task to execute its command/script

install_crate: Option<InstallCrate>

if defined, the provided crate will be installed (if needed) before running the task

install_crate_args: Option<Vec<String>>

additional cargo install arguments

install_script: Option<Vec<String>>

if defined, the provided script will be executed before running the task

command: Option<String>

The command to execute

args: Option<Vec<String>>

The command args

script: Option<ScriptValue>

If command is not defined, and script is defined, the provided script will be executed

script_runner: Option<String>

The script runner (defaults to cmd in windows and sh for other platforms)

script_extension: Option<String>

The script file extension

run_task: Option<RunTaskInfo>

The task name to execute

dependencies: Option<Vec<String>>

A list of tasks to execute before this task

toolchain: Option<String>

The rust toolchain used to invoke the command or install the needed crates/components

Methods

impl PlatformOverrideTask

pub fn extend(self: &mut PlatformOverrideTask, task: &mut Task)

Copies values from the task into self.

Arguments

task - The task to copy from