Crate cli

A collection of tools, mostly related to importing maps and scenarios. These are bundled as a single executable to reduce the cost of static linking in the release’s file size.

Modules

augment_scenario

clip_osm

generate_houses

geojson_to_osmosis

import_grid2demand

import_scenario

one_step_import

pick_geofabrik

Enums

Command

Functions

dump_json

import_json_map

main

minify_map

random_scenario