Crate cli

A collection of tools, mostly related to importing maps and scenarios. These are bundled as a single executable to reduce the cost of static linking in the release’s file size.

Modules

augment_scenario 🔒

clip_osm 🔒

generate_houses 🔒

import_grid2demand 🔒

import_scenario 🔒

one_step_import 🔒

Enums

Command 🔒

Functions

dump_json 🔒

import_json_map 🔒

main 🔒

minify_map 🔒

parse_modifiers 🔒

prebake_scenario 🔒

random_scenario 🔒

regenerate_everything_externally 🔒

Type Aliases

ModifierList 🔒