Crate cliExperimental

A toolkit for building command-line programs in Rust.

cli aims to

reduce boilerplate

canonicalize output formatting

encourage adherence to the conventions of standard streams (e.g. when to print to stderr vs stdout)

encourage appropriate use of exit statuses

Concretely, cli is a collection of simple functions for formatting command-line output, defining flags, and (eventually) things like integrating with sysexits.

It uses getopts for argument parsing and thus currently inherits getopt's experimental stability level.

Functions

argv0
The name by which the executed program was called.

helpopt
Create a help flag OptGroup.

parse_args
Parse the command-line arguments with which the program was executed according to a collection of OptGroups.

usage_string
Construct a canonical usage string from a collection of OptGroups.

version_string
Construct a version string.

versionopt
Create a version flag OptGroup.