Crate cli

Cli is a Rust crate for creating beautiful command line applications.

Reexports

pub use term::{Black, Red, Green, Yellow, Blue, Magenta, Cyan, White};

Structs

Editor
One editor for you to edit the given text or file.

ProgressBar
Showing a progress bar. Examples:

Style
Styles a text with ANSI styles. This styling is self contained which means that at the end of the string a reset code is issued. Examples:

Enums

Color

Functions

clear
Clears the terminal screen.

confirm
Prompts for confirmation (yes/no question).

get_terminal_size
Returns the current size of the terminal in the form (width, height) in columns and rows, usage example:

isatty
Check output device is a terminal or not.

print_via_pager
Show text via an pager.

prompt
Prompts a user for input.

sprintln
Styles the message with ANSI styles and println it.