Crate cli

Modules

bench 🔒

database 🔒

error

export 🔒

import 🔒

meta_snapshot 🔒

Structs

AttachCommand 🔒

BenchTableMetadataCommand

DatabaseClient

ExportCommand

ImportCommand

MetaRestoreCommand

Restore metadata snapshot tool. This tool is used to restore metadata snapshot from etcd, pg or mysql. It will restore the metadata snapshot from local file or s3 bucket.

MetaSnapshotCommand

Export metadata snapshot tool. This tool is used to export metadata snapshot from etcd, pg or mysql. It will dump the metadata snapshot to local file or s3 bucket. The snapshot file will be in binary format.

Traits

Tool