Crate cli

The CLI binary file is intended to be used by non-rustcean developers to interact with the rust-csv library. Specifically written for Data Scientist and Data Engineers using high-level programming languages such as Python or R. The CLI tool will allow users to read csv files and perform filtering operations, such as filtering by a specific column and query. This simple tool will reduct the I/O overhead of reading csv files and filtering data, reducing the time spent on data preprocessing.

Example

cli transform \
 --query "'1'" \
 --column "val" \
 --output-path "./test.csv" \
 filter \
 --file "../sdk/test/example/data.csv"

Errors

This function will return an error if the command is not recognized or if there is an issue with the filtering operations.

Enums

Cli 🔒

Cli enum to hold the different commands

Functions

main 🔒

Main entry point for the CLI tool to read csv files and perform filtering operations