Crate cli

cargo-make

Rust task runner and build tool.
The cargo-make task runner enables to define and configure sets of tasks and run them as a flow.
A task is a command or a script to execute.
Tasks can have dependencies which are also tasks that will be executed before the task itself.
With a simple toml based configuration file, you can define a multi platform build script that can run build, test, documentation generation, bench tests execution, security validations and more by running a single command.

Installation

In order to install, just run the following command

cargo install cargo-make

This will install cargo-make in your ~/.cargo/bin.
Make sure to add ~/.cargo/bin directory to your PATH variable.

Contributing

See contributing guide

License

Developed by Sagie Gur-Ari and licensed under the Apache 2 open source license.

Modules

cli

cli

cli_commands

cli_commands

cli_parser

cli_parser

config

config

error

logger

log

runner

runner

types

types

Functions

run_cli

Handles the command line arguments and executes the runner.