Module cli::commands

Command struct to interact with client.

Traits

Command
Trait to perform client operations.

Functions

blocking_cmd
Check whether a command is blocking.

debug_format_cmd
Check whether a command is debugging command.

get_commands
Returns all the commands available, as well as the reverse index from the aliases to the commands.

is_address
Check whether the input string is a valid libra address.

is_authentication_key
Check whether the input string is a valid libra authentication key.

parse_cmd
Parse a cmd string, the first element in the returned vector is the command to run

print_subcommand_help
Print the help message for all sub commands.

report_error
Print the error and bump up error counter.

subcommand_execute
Execute sub command.