Module cli::types

types

Defines the various types and aliases used by cargo-make.

Structs

Cache
Holds persisted data used by cargo-make

CliArgs
Holds CLI args

CommandSpec
Command info

Config
Holds the entire configuration such as task definitions and env vars

ConfigSection
Holds the configuration found in the makefile toml config section.

CrateDependencyInfo
Holds crate dependency info.

CrateInfo
Holds crate information loaded from the Cargo.toml file.

EnvInfo
Holds env information

EnvValueInfo
Holds a single task configuration such as command and dependencies list

ExecutionPlan
Execution plan which defines all steps to run and the order to run them

ExternalConfig
Holds the entire externally read configuration such as task definitions and env vars where all values are optional

FlowInfo
Holds flow information

GitInfo
Holds git info for the given repo directory

GlobalConfig
Holds configuration info for cargo-make

InstallCrateInfo
Holds instructions how to install the crate

InstallRustupComponentInfo
Holds instructions how to install a rustup component

PackageInfo
Holds crate package information loaded from the Cargo.toml file package section.

PlatformOverrideTask
Holds a single task configuration for a specific platform as an override of another task

RustVersionCondition
Rust version condition structure

Step
Execution plan step to execute

Task
Holds a single task configuration such as command and dependencies list

TaskCondition
Holds condition attributes

Workspace
Holds crate workspace info, see http://doc.crates.io/manifest.html#the-workspace-section

Enums

CrateDependency
Holds crate dependency info.

EnvValue
Holds the env value or script

InstallCrate
Install crate name or params

Functions

get_platform_name
Returns the platform name