Module cli::types

types

Defines the various types and aliases used by cargo-make.

Structs

Cache
Holds persisted data used by cargo-make

CargoMetadata
The output of cargo metadata command

CliArgs
Holds CLI args

CommandSpec
Command info

Config
Holds the entire configuration such as task definitions and env vars

ConfigSection
Holds the configuration found in the makefile toml config section.

CrateDependencyInfo
Holds crate dependency info.

CrateInfo
Holds crate information loaded from the Cargo.toml file.

EnvFileInfo
Env file path and attributes

EnvInfo
Holds env information

EnvValueConditioned
Env value provided by decoding other values

EnvValueDecode
Env value provided by decoding other values

EnvValueScript
Env value provided by a script

EnvValueUnset
Enables to unset env variables

ExecutionPlan
Execution plan which defines all steps to run and the order to run them

ExtendOptions
Extend with more fine tuning options

ExternalConfig
Holds the entire externally read configuration such as task definitions and env vars where all values are optional

FileScriptValue
Script file name

FlowInfo
Holds flow information

FlowState
Holds mutable flow state

GlobalConfig
Holds configuration info for cargo-make

InstallCargoPluginInfo
Holds instructions how to install the cargo plugin

InstallCrateInfo
Holds instructions how to install the crate

InstallRustupComponentInfo
Holds instructions how to install a rustup component

ModifyConfig
Holds properties to modify the core tasks

PackageInfo
Holds crate package information loaded from the Cargo.toml file package section.

PlatformOverrideTask
Holds a single task configuration for a specific platform as an override of another task

RunTaskDetails
Holds the run task information

RunTaskRoutingInfo
Holds the run task routing information

RustVersionCondition
Rust version condition structure

ScriptSections
Script content split to parts to enable a more fine tuned extension capability

Step
Execution plan step to execute

Task
Holds a single task configuration such as command and dependencies list

TaskCondition
Holds condition attributes

TaskIdentifier
An identifier for a task

TestArg
Arguments used to check whether a crate or rustup component is installed.

WatchOptions
Holds watch options

Workspace
Holds crate workspace info, see http://doc.crates.io/manifest.html#the-workspace-section

Enums

CrateDependency
Holds crate dependency info.

DependencyIdentifier
A dependency, defined either as a string or as a Dependency object

DeprecationInfo
Holds deprecation info such as true/false/message

EnvFile
Holds the env file path and attributes

EnvValue
Holds the env value or script

Extend
Holds makefile extend value

InstallCrate
Install crate name or params

RunTaskInfo
Run task info

RunTaskName
Holds the run task name/s

ScriptValue
Script value (text, file name, ...)

TaskWatchOptions
Holds watch options or simple true/false value

Functions

get_platform_name
Returns the platform name