Struct cli::types::CliArgs

pub struct CliArgs {

    pub command: String,
    pub build_file: Option<String>,
    pub task: String,
    pub profile: Option<String>,
    pub log_level: String,
    pub disable_color: bool,
    pub cwd: Option<String>,
    pub env: Option<Vec<String>>,
    pub env_file: Option<String>,
    pub disable_workspace: bool,
    pub disable_on_error: bool,
    pub allow_private: bool,
    pub skip_init_end_tasks: bool,
    pub skip_tasks_pattern: Option<String>,
    pub print_only: bool,
    pub list_all_steps: bool,
    pub list_category_steps: Option<String>,
    pub diff_execution_plan: bool,
    pub disable_check_for_updates: bool,
    pub experimental: bool,
    pub arguments: Option<Vec<String>>,
    pub output_format: String,
    pub output_file: Option<String>,
    pub print_time_summary: bool,

}

Holds CLI args

Fields
command: String

The command name

build_file: Option<String>

The external Makefile.toml path

task: String

The task to invoke

profile: Option<String>

The profile name

log_level: String

Log level name

disable_color: bool

Disables colorful output

cwd: Option<String>

Current working directory

env: Option<Vec<String>>

Environment variables

env_file: Option<String>

Environment variables file

disable_workspace: bool

Prevent workspace support

disable_on_error: bool

Prevent on error flow even if defined in config section

allow_private: bool

Allow invocation of private tasks

skip_init_end_tasks: bool

If true, the init and end tasks are skipped

skip_tasks_pattern: Option<String>

Skip tasks that match the provided pattern

print_only: bool

Only print the execution plan

list_all_steps: bool

List all known steps

list_category_steps: Option<String>

List steps for a given category

diff_execution_plan: bool

Diff flows

disable_check_for_updates: bool

Disables the update check during startup

experimental: bool

Allows access unsupported experimental predefined tasks

arguments: Option<Vec<String>>

additional command line arguments

output_format: String

Output format

output_file: Option<String>

Output file name

print_time_summary: bool

Print time summary at end of the flow

Implementations

impl CliArgs

pub fn new() -> CliArgs

Creates and returns a new instance.