Struct cli::types::ConfigSection

pub struct ConfigSection {
    pub skip_core_tasks: Option<bool>,
    pub modify_core_tasks: Option<ModifyConfig>,
    pub init_task: Option<String>,
    pub end_task: Option<String>,
    pub on_error_task: Option<String>,
    pub additional_profiles: Option<Vec<String>>,
    pub min_version: Option<String>,
    pub default_to_workspace: Option<bool>,
    pub load_script: Option<Vec<String>>,
    pub linux_load_script: Option<Vec<String>>,
    pub windows_load_script: Option<Vec<String>>,
    pub mac_load_script: Option<Vec<String>>,
}

Holds the configuration found in the makefile toml config section.

Fields
skip_core_tasks: Option<bool>

If true, the default core tasks will not be loaded

modify_core_tasks: Option<ModifyConfig>

Modify core tasks config

init_task: Option<String>

Init task name which will be invoked at the start of every run

end_task: Option<String>

End task name which will be invoked at the end of every run

on_error_task: Option<String>

The name of the task to run in case of any error during the invocation of the flow

additional_profiles: Option<Vec<String>>

Additional profile names to load

min_version: Option<String>

Minimum cargo-make/makers version

default_to_workspace: Option<bool>

The task.workspace default value

load_script: Option<Vec<String>>

Invoked while loading the descriptor file but before loading any extended descriptor

linux_load_script: Option<Vec<String>>

acts like load_script if runtime OS is Linux (takes precedence over load_script)

windows_load_script: Option<Vec<String>>

acts like load_script if runtime OS is Windows (takes precedence over load_script)

mac_load_script: Option<Vec<String>>

acts like load_script if runtime OS is Mac (takes precedence over load_script)

Methods

impl ConfigSection

pub fn new() -> ConfigSection

Creates and returns a new instance.

pub fn apply(self: &mut ConfigSection, modify_config: &ModifyConfig)

Apply modifications

pub fn extend(self: &mut ConfigSection, extended: &mut ConfigSection)

Copies values from the config section into self.

Arguments

task - The task to copy from

pub fn get_load_script(self: &ConfigSection) -> Option<Vec<String>>

Returns the load script based on the current platform